refactor: moved get_data_set_from_config from utils.cdf.datasets_calls into utils.retrieve
